examples/requirements.txt
mysql-connector-python==8.0.19
pytest==3.2.1
dnspython==1.16.0
